metrics_api/cdk/requirements.txt
aws-cdk-lib==2.173.4
constructs>=10.0.0,<11.0.0